Move to integration/ and move repositories to individual files
