+++ integration/vehicle-inspection/.gitignore
