+++ integration/vehicle-inspection/src/index.ts
