+++ integration/vehicle-inspection/src/server/index.ts
@@ -14,52 +14,6 @@ import * as express from 'express';
 import {cloneDeep, filter, isFunction, map} from 'lodash';
 import {VehicleInspectionTask} from '../vehicle-inspection/VehicleInspectionTask';
 
-class InMemoryTaskRepository implements TaskRepository {
-  private data: Task[] = [];
-  public createBatch(tasks: Task[]) {
-    this.data = this.data.concat(tasks);
-    return Promise.resolve(this.data);
-  }
-}
-
-class InMemoryProcessRepository implements ProcessRepository {
-  private data: Process[] = [];
-  constructor(protected seedData: Process[], protected taskRepository: TaskRepository) {
-    this.data = cloneDeep(seedData);
-  }
-  public getAll() {
-    return Promise.resolve(this.data);
-  }
-
-  public create(process: Process) {
-    return this.taskRepository.createBatch(process.tasks)
-    .then(tasks => {
-      process.tasks = tasks;
-      this.data.push(process);
-      return process;
-    });
-  }
-}
-
-class InMemoryProcessInstanceRepository implements ProcessInstanceRepository {
-  private data: ProcessInstance[];
-  constructor(protected seedData: ProcessInstance[]) {
-    this.data = cloneDeep(seedData);
-  }
-  public getAll() {
-    return Promise.resolve(this.data);
-  }
-
-  public create(process: ProcessInstance) {
-    this.data.push(process);
-    return Promise.resolve(process);
-  }
-
-  public getUnassigned() {
-    return Promise.resolve(filter(this.data, p => !p.assigneeId));
-  }
-}
-
 class Server {
   public app = express();
   public port = process.env.PORT || 3000;
