+++ integration/vehicle-inspection/src/vehicle-inspection/VehicleInspectionFormResult.ts
