+++ integration/vehicle-inspection/src/vehicle-inspection/VehicleInspectionTask.ts
