+++ integration/vehicle-inspection/tsconfig.json
