+++ lerna.json
@@ -3,7 +3,7 @@
   "packages": [
     "cloud/*",
     "client/*",
-    "examples/*",
+    "integration/*",
     "demo/*"
   ],
   "version": "independent"
